Added Features
I resized the window, changed the text, and made the window borderless. I also changed the font and increased the size of the text.

I have also implemented a feature where the clock now shows the date.

Since the window is now borderless, you need to right click for a menu to pop up, allowing you to close it.